Feat: 코디 요청 응답 기능 구현

+++ app/src/main/java/com/umc/upstyle/data/network/Apiservice.kt
@@ -18,7 +18,7 @@ interface ApiService {
     // 특정 사용자 옷장 조회 API
     @GET("closets/")
     fun getUserCloset(
-        @Query("userId") userId: Long
+        @Query("userId") userId: Int
     ): Call<ClosetResponse>
 
     // 옷장 카테고리별 조회 API
